revert to regex, rename to url
