+++ policy/option.go
@@ -20,7 +20,7 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	urlutil "github.com/yahoojapan/athenz-authorizer/internal"
+	urlutil "github.com/yahoojapan/athenz-authorizer/internal/url"
 	"github.com/yahoojapan/athenz-authorizer/pubkey"
 )
